feat(invoice): invoiceListの責務分離リファクタリング
Issue #23 に基づき、請求書一覧画面 (invoiceList) のJavaScriptをリファクタリング。

- データ取得ロジックを `invoiceService.js` に分離。
- UI描画ロジックを `invoiceRenderer.js` に分離。
- `invoiceList.js` は、サービスとレンダラーを呼び出す役割に限定。

これにより、関心の分離が徹底され、コードの保守性とテスト容易性が向上しました。
